@@ -32,12 +32,12 @@
         }
     ],
     "require": {
-        "php": ">=7.4",
-        "illuminate/events": "~7.0 || ~8.0 || ~9.0",
-        "illuminate/notifications": "~7.0 || ~8.0 || ~9.0",
-        "illuminate/queue": "~7.0 || ~8.0 || ~9.0",
-        "illuminate/support": "~7.0 || ~8.0 || ~9.0",
-        "pusher/pusher-push-notifications": "^1.1"
+        "php": ">=8.0",
+        "illuminate/events": "~8.0 || ~9.0",
+        "illuminate/notifications": "~8.0 || ~9.0",
+        "illuminate/queue": "~8.0 || ~9.0",
+        "illuminate/support": "~8.0 || ~9.0",
+        "pusher/pusher-push-notifications": "^2.0"
     },
     "require-dev": {
         "mockery/mockery": "^1.3",
